fixed type def issues, tsconfig, updated deps
